Add exam sentence
