+++ Karasuma/Karasuma.Shared/Helper/Extension.cs
@@ -21,4 +21,37 @@ namespace Karasuma.Helper
             }
         }
     }
+
+    public static class Combinations<T>
+    {
+        public static List<T[]> GetCombinations(List<T[]> sourceList)
+        {
+            List<T[]> resultList = new List<T[]>();
+            Stack<T> stack = new Stack<T>();
+            GetCombinationsCore(stack, resultList, sourceList);
+
+            return resultList;
+        }
+
+        private static void GetCombinationsCore(Stack<T> stack, List<T[]> resultList, List<T[]> sourceList)
+        {
+            int dimension = stack.Count;
+            if (sourceList.Count <= dimension)
+            {
+                T[] array = stack.ToArray();
+                Array.Reverse(array);
+                resultList.Add(array);
+                return;
+            }
+            else
+            {
+                foreach (T item in sourceList[dimension])
+                {
+                    stack.Push(item);
+                    GetCombinationsCore(stack, resultList, sourceList);
+                    stack.Pop();
+                }
+            }
+        }
+    }
 }
