+++ Karasuma/Karasuma.Shared/Models/Chunk.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Karasuma.Helper;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Diagnostics;
@@ -491,39 +492,4 @@ namespace Karasuma
 
         
     }
-
-    
-    public static class Combinations<T>
-    {
-        public static List<T[]> GetCombinations(List<T[]> sourceList)
-        {
-            List<T[]> resultList = new List<T[]>();
-            Stack<T> stack = new Stack<T>();
-            GetCombinationsCore(stack, resultList, sourceList);
-
-            return resultList;
-        }
-
-        private static void GetCombinationsCore(Stack<T> stack, List<T[]> resultList, List<T[]> sourceList)
-        {
-            int dimension = stack.Count;
-            if (sourceList.Count <= dimension)
-            {
-                T[] array = stack.ToArray();
-                Array.Reverse(array);
-                resultList.Add(array);
-                return;
-            }
-            else
-            {
-                foreach (T item in sourceList[dimension])
-                {
-                    stack.Push(item);
-                    GetCombinationsCore(stack, resultList, sourceList);
-                    stack.Pop();
-                }
-            }
-        }
-    }
-
 }
