+++ Karasuma/Karasuma.Wasm/WasmScripts/AppManifest.js
@@ -1,7 +1,6 @@
 ﻿var UnoAppManifest = {
 
-    splashScreenImage: "Assets/SplashScreen.png",
-    splashScreenColor: "#0078D7",
-    displayName: "Karasuma"
-
+    //splashScreenImage: "Assets/SplashScreen.png",
+    //splashScreenColor: "#0078D7",
+    displayName: "uno typing"
 }
